+++ src/apitypes/graphql/graphql.document.ts
@@ -62,19 +62,3 @@ export const buildGraphQLDocument = async (parsedFile: TextFile, file: BuildConf
 export const dumpGraphQLDocument: DocumentDumper<GraphApiSchema> = (document) => {
   return new Blob([printGraphApi(document.data)], { type: 'text/plain' })
 }
-
-export async function createGraphQLExportDocument(
-  filename: string,
-  data: string,
-  format: ExportFormat,
-): Promise<ExportDocument> {
-  if (format !== GRAPHQL_API_TYPE) {
-    throw new Error('Unsupported format type')
-  }
-  const exportFilename = `${getDocumentTitle(filename)}.${GRAPHQL_API_TYPE}`
-
-  return {
-    data: new Blob([data], { type: 'application/graphql' }),
-    filename: exportFilename,
-  }
-}
